backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
dj-database-url==2.2.0
supabase==2.10.0
